% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_ei_results.R
\name{ei_cv_summary}
\alias{ei_cv_summary}
\title{Obtain cell value estimates from ssEI ecological inference models.}
\usage{
ei_cv_summary(
  ei_stanfit,
  pars = "cell_values",
  probs = c(0.025, 0.25, 0.5, 0.75, 0.975),
  ...
)
}
\arguments{
\item{ei_stanfit}{Usually the result of running the \code{ei_estimate} command. A \code{stanfit} object that results from running the one of the ssEI models.}

\item{pars}{the name of the parameters to extract (usually \code{cell_values})}

\item{probs}{probabilities to pass to rstan::summary.stanfit}

\item{...}{}
}
\value{
A tibble with estimates of cell values for each combination of area number, row number and column number.
}
\description{
Obtain cell value estimates from ssEI ecological inference models.
}
