% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_ei_results.R
\name{ei_overall_row_rate_summary}
\alias{ei_overall_row_rate_summary}
\title{Obtain overall row rate estimates (total across all areas) from an ssEI model.}
\usage{
ei_overall_row_rate_summary(
  ei_stanfit,
  pars = "overall_row_rates",
  probs = c(0.025, 0.25, 0.5, 0.75, 0.975),
  ...
)
}
\arguments{
\item{ei_stanfit}{Usually the result of running the \code{ei_estimate} command. A \code{stanfit} object that results from running the one of the ssEI models.}

\item{pars}{the name of the parameters to extract (usually \code{overall_row_rates})}

\item{probs}{probabilities to pass to rstan::summary.stanfit}

\item{...}{}
}
\value{
A tibble with estimates of row rates (the percentage of the row total in each column) for each combination of row number and column number (summary total across all areas). For rows which sum to zero (and hence consist only of zero cells) a value of -1 is returned.
}
\description{
Obtain overall row rate estimates (total across all areas) from an ssEI model.
}
