% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssEI-package.R
\docType{package}
\name{ssEI-package}
\alias{ssEI-package}
\alias{ssEI}
\title{The 'ssEI' package.}
\description{
Ecological inference using sequential sampling methods
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.32.3. https://mc-stan.org
}
